+++ packages/client/test/own-common.test.js
@@ -8,10 +8,10 @@ const ownWrapper = require('./own-common/helpers/own-wrapper-test');
 const syncTests = require('./own-common/helpers/sync-test');
 const eventsTests = require('./own-common/helpers/events-test');
 const localStorageTests = require('./own-common/helpers/local-storage-test');
+const asyncStorageTests = require('./own-common/helpers/async-storage-test');
 const restTests = require('./own-common/helpers/rest-test');
 const socketioTests = require('./own-common/helpers/socket-io-test');
-let LocalStorage = require('node-localstorage').LocalStorage;
-global.localStorage = new LocalStorage('./.scratch');
+
 
 const OwnClass = require('../src/own-common');
 
@@ -67,7 +67,8 @@ describe(`${package}Wrapper tests`, () => {
   ownWrapper(`${package}Wrapper specific functionality`, app, errors, ownclassWrapper, 'ownWrapper', verbose, true);
   syncTests(`${package}Wrapper sync functionality`, app, errors, init, 'syncTests', verbose, 9100, true);
   eventsTests(`${package}Wrapper events functionality`, app, errors, ownclassWrapper, 'wrapperEvents', verbose);
-  localStorageTests(`${package}Wrapper storage functionality`, app, errors, ownclassWrapper, 'wrapperStorage', verbose);
+  localStorageTests(`${package}Wrapper localStorage functionality`, app, errors, ownclassWrapper, 'wrapperLocalStorage', verbose);
+  asyncStorageTests(`${package}Wrapper asyncStorage functionality`, app, errors, ownclassWrapper, 'wrapperAsyncStorage', verbose);
   restTests(`${package}Wrapper REST functionality`, app, errors, ownclassWrapper, 'wrapperREST', verbose, 7886, true);
   socketioTests(`${package}Wrapper socket.io functionality`, app, errors, ownclassWrapper, 'wrapperSocketIo', verbose, 7886, true);