+++ packages/server/test/common.test.js
@@ -85,13 +85,13 @@ describe('Common tests - test of common functions', () => {
         it('short uuid', () => {
           let uuid1 = genUuid(true);
           let uuid2 = genUuid(true);
-          expect(uuid1.length).to.satisfy((v) => {return v<16}, 'short uuid length is wrong');
+          expect(uuid1.length).to.satisfy((v) => {return v<11}, 'short uuid length is wrong');
           expect(uuid1).to.not.equal(uuid2, 'the two uuids are identical');
         });
         it('long uuid', () => {
           let uuid1 = genUuid(false);
           let uuid2 = genUuid(false);
-            expect(uuid1.length).to.equal(36, 'long uuid length is wrong');
+            expect(uuid1.length).to.equal(21, 'long uuid length is wrong');
           expect(uuid1).to.not.equal(uuid2, 'the two uuids are identical');
         });
       });