+++ packages/server/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 
+## [2.0.0](http://github.com/feathersjs-offline/owndata-ownnet/packages/server/compare/v1.1.1...v2.0.0) (2021-10-25)
+
+Code clean-up and support for `dates` option ensuring ISO-formatted strings in
+results sets always are returned as Date's.
+
+Version bumped to match @feathersjs-offline/client.
 ## [1.1.1](http://github.com/feathersjs-offline/owndata-ownnet/packages/server/compare/v1.1.0...v1.1.1) (2021-08-01)
 
 **Note:** Version bump only for package @feathersjs-offline/server